requirements.txt
Django==5.2.8
aiogram==3.22.0
python-decouple==3.8
pillow==11.1.0
requests==2.32.5
psycopg2-binary==2.9.11
openpyxl==3.1.5